feat: add string commands INCR/DECR/INCRBY/DECRBY/MGET/MSET/SETNX/SETEX/APPEND/STRLEN
Atomic increment/append use ConcurrentHashMap.compute(). CommandResult.Integer
widened to long to handle counters beyond Integer.MAX_VALUE.

+++ src/main/java/org/valarpirai/redis/command/CommandResult.java
@@ -17,7 +17,7 @@ public sealed interface CommandResult
 
   record Bulk(String value) implements CommandResult {}
 
-  record Integer(int value) implements CommandResult {}
+  record Integer(long value) implements CommandResult {}
 
   record Error(String message) implements CommandResult {}
 
@@ -37,7 +37,7 @@ public sealed interface CommandResult
     return new Bulk(value);
   }
 
-  static CommandResult integer(int value) {
+  static CommandResult integer(long value) {
     return new Integer(value);
   }
